Fixed gitCheckout method and extended App to do setp by step commits.
gitCheckout was buggy because of internal issues of the JGit API. Work around for this problem was using the runtime execution to use the git shell commands.
Also a clean --force and a reset --hard is necessary before every new checkout. Now the Mehtods behaves as expected.

App was extended to now check out every commit on its own and continues to link the changes only to features in the current commit.
Commit Messages stay the same, but the file committed to ecco changes accordingly.