Changed Feature Structure to work with defines for COAN
Later on COAN should be used to preprocess the c file before committing. To achieve this it was necessary to link a definition to a feature name. If a feature has multiple names ( because of a boolean expression) each name needs its own definition.
For this the ParsedCondition class was added.

Now the definition is parsed and stored in the feature. Also the preprocess Method in the FeaturePreprocessor now builds a coan command.
Next step is to use coan to preprocess.